requirements.txt
# DataPulse — Python dependencies
# Install with: pip install -r requirements.txt

# ─── Data processing ──────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.13.0
openpyxl>=3.1.2
xlrd>=2.0.1

# ─── Visualisation ────────────────────────────────────────────────────────────
plotly>=5.22.0

# ─── API framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
asyncpg>=0.29.0
greenlet>=3.0.0,<4.0

# ─── Cache ────────────────────────────────────────────────────────────────────
redis>=5.0.0

# ─── Auth ─────────────────────────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-jose[cryptography]>=3.3.0

# ─── Validation ───────────────────────────────────────────────────────────────
pydantic>=2.7.0
email-validator>=2.1.0.post1

# ─── LLM ──────────────────────────────────────────────────────────────────────
groq>=0.9.0

# ─── Config ───────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

google-auth>=2.29.0
requests>=2.31.0
